feat: 댓글 입력/삭제 시 Article의 commentCnt 증가/감소 구현

+++ backend/src/main/java/com/example/backend/service/CommentService.java
@@ -7,6 +7,7 @@ import com.example.backend.dto.CommentDto;
 import com.example.backend.dto.Comments;
 import com.example.backend.dto.request.CommentCreateRequest;
 import com.example.backend.dto.request.CommentUpdateRequest;
+import com.example.backend.model.Article;
 import com.example.backend.model.Comment;
 import com.example.backend.repository.ArticleRepository;
 import com.example.backend.repository.CommentRepository;
@@ -21,6 +22,8 @@ import org.springframework.stereotype.Service;
 @Service
 public class CommentService {
 
+    private static final int COUNT = 1;
+
     private final ArticleRepository articleRepository;
     private final CommentRepository commentRepository;
 
@@ -32,14 +35,25 @@ public class CommentService {
     }
 
     public CommentDto create(CommentCreateRequest commentCreateRequest, Long articleId, Long currentUserId) {
+        Article article = articleRepository.findById(articleId)
+            .orElseThrow(() -> {
+                throw new CustomException(ResponseCode.POST_NOT_FOUND);
+            });
+
         Comment comment = commentCreateRequest.toEntity(articleId, currentUserId);
         commentRepository.save(comment);
+
+        article.setCommentCnt(article.getCommentCnt() + COUNT);
+        articleRepository.save(article);
 
         return CommentDto.from(comment);
     }
 
     public CommentDto update(CommentUpdateRequest commentUpdateRequest, Long articleId, Long currentUserId, Long commentId) {
-        validateArticle(articleId);
+        articleRepository.findById(articleId)
+            .orElseThrow(() -> {
+                throw new CustomException(ResponseCode.POST_NOT_FOUND);
+            });
         validateAuth(commentId, currentUserId);
 
         Comment updateComment = commentRepository.findById(commentId)
@@ -52,12 +66,22 @@ public class CommentService {
     }
 
     public HttpStatus delete(Long articleId, Long commentId) {
-        validateArticle(articleId);
+        Article article = articleRepository.findById(articleId)
+            .orElseThrow(() -> {
+                throw new CustomException(ResponseCode.POST_NOT_FOUND);
+            });
+
+        if (article.getCommentCnt() < 0) {
+            throw new CustomException(ResponseCode.COMMENT_NOT_FOUND);
+        }
 
         Comment deleteComment = commentRepository.findById(commentId)
             .orElseThrow(() -> {throw new CustomException(ResponseCode.COMMENT_NOT_FOUND);});
 
         commentRepository.deleteById(deleteComment.getCommentId());
+
+        article.setCommentCnt(article.getCommentCnt() - COUNT);
+        articleRepository.save(article);
 
         return HttpStatus.OK;
     }
@@ -67,13 +91,6 @@ public class CommentService {
 
         return Comments.builder()
             .comments(comments).build();
-    }
-
-    private void validateArticle(Long articleId) {
-        articleRepository.findById(articleId)
-            .orElseThrow(() -> {
-                throw new CustomException(ResponseCode.POST_NOT_FOUND);
-            });
     }
 
     private void validateAuth(Long commentId, Long currentUserId) {
